improve agent resilliance
In short, even if the device file does not exist or the forge
platform is offline, keep retrying. This means the administrator
does not have to fire up the device again after forge is online.

Also, the checking of  file permission to ensure we check everything
is in order before calling the POST to provision.