fix: Use rename over copy + delete
In my previous commit, the choice was made to move the config file and
afterwards delete the original. In situations where a Network drive is
used, this is better.

This change moves it to renaming, which is an atomic operation for
nearly all file systems, just not networked ones. However, it's superior
for simplicty reasons and the error mode of NFS are unlikely to show up
here either way.